refactor: move errors to kv package
